requirements.txt
fastapi 
uvicorn[standard]
pandas 
numpy 
scikit-learn 
tensorflow 
keras 
paho-mqtt 
kafka-python 
psycopg2-binary 
sqlalchemy 
langchain 
langchain-community 
chromadb 
sentence-transformers 
pydantic 
python-dotenv 
matplotlib 
seaborn 
plotly 
streamlit 
pytest 
httpx